Add files via upload
SQLi stands for SQL Injection, which is a type of cyber attack that targets the vulnerabilities in a web application's database layer. It occurs when an attacker is able to manipulate or inject malicious SQL (Structured Query Language) code into input fields or parameters that are used in database queries.

The primary goal of SQL injection is to manipulate the application's SQL query in such a way that it allows unauthorized access to the database or the ability to perform unintended actions on the database. This can lead to unauthorized access, data theft, data manipulation, and in some cases, full control over the affected system.